# api/openapi.yml
openapi: 3.0.1
info:
  title: Alcarin API
  description: Alcarin api
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
...
  '/users/current/characters/{characterId}':
    get:
      summary: One user character
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CharacterModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
      operationId: 'get-users-current-characters-:characterId'
      description: Getting data model for character
      security:
        - Authorization: []
    parameters:
      - schema:
          type: string
        name: characterId
        in: path
        description: Character ID
        required: true
components:
  schemas:
...
    CharacterModel:
      type: object
      title: CharacterModel
      additionalProperties: false
      description: Model of race data
      properties:
        name:
          type: string
        id:
          type: string
        race:
          type: string
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer
      description: Token for current user