# api/src/server/rest/controllers/character.ctrl.ts
import { AppRequestHandler } from 'express';
import boom from '@hapi/boom';
import status from 'http-status-codes';
import { logger } from 'src/server/core/helpers/logger';
import {
  createCharacter,
  getCharacters,
} from 'src/server/services/account-access.service';
import { AvailableRace } from 'src/server/plugins/game/races/available-race-provider';
import { Character } from 'src/domain/game/character/character';

interface CreateNewCharacterReq {
  body: {
    name: string;
    race: AvailableRace;
  };
}

interface OneCharacterPayload {
  params: {
    characterId: string;
  };
}

export const createNewCharacter: AppRequestHandler<CreateNewCharacterReq> = async (
  req,
  res
) => {
  const { name, race } = req.body;
  const accountId = req.accountId;

  try {
    const character = await createCharacter(accountId, name, race);
    logger.info(`Character "${name} with race ${race}" created`);

    return res.status(status.CREATED).send(mapCharacterToResponse(character));
  } catch (err) {
    throw boom.badRequest(err.message);
  }
};

export const findCharactersForUsers: AppRequestHandler = async (req, res) => {
  const accountId = req.accountId;

  try {
    const characters = await getCharacters(accountId);

    return res.status(status.OK).send(characters.map(mapCharacterToResponse));
  } catch (err) {
    throw boom.badRequest(err.message);
  }
};

function mapCharacterToResponse(character: Character<AvailableRace>) {
  return {
    name: character.name,
    id: character.id,
    race: character.raceKey,
  };
}